+++ app.js
@@ -1,16 +1,18 @@
 'use strict';
 
+var request = require('request');
+var querystring = require('querystring');
+
 var restify = require('restify');
 var botbuilder = require('botbuilder');
+var config = require('nconf').env().argv().file({ file: './localConfig.json' });
 
 function createConsoleConnector() {
-  console.log('Hi.. ask me a question!');
+  console.log('Hi.. ask about the Microsoft Bot Framework and I\'ll do my best to answer.');
   return new botbuilder.ConsoleConnector().listen();
 }
 
 function createChatConnector() {
-
-  var config = require('nconf').env().file({ file: './localConfig.json' });
 
   var opts = {
     appId: config.get('MICROSOFT_APP_ID'),
@@ -19,7 +21,6 @@ function createChatConnector() {
 
   var chatConnector = new botbuilder.ChatConnector(opts);
 
-  // Create http server
   var server = restify.createServer();
   server.listen(process.env.port || process.env.PORT || 3978, function () {
     console.log('%s listening to %s', server.name, server.url);
@@ -30,9 +31,38 @@ function createChatConnector() {
   return chatConnector;
 }
 
+function qna(q, cb) {
+  q = querystring.escape(q);
+  request(config.get('qnaUri') + q, function (error, response, body) {
+    if (error) {
+      cb(error, null);
+    }
+    else if (response.statusCode !== 200) {
+      cb(response, null);
+    }
+    else {
+      cb(null, body);
+    }
+  });
+}
+
 function main() {
-  var connector = createConsoleConnector();
-  var bot = require('./lib/qnabot_minimal.js')(connector);
+  var connector = config.get('console') ? createConsoleConnector() : createChatConnector();
+  var bot = new botbuilder.UniversalBot(connector);
+
+  bot.dialog('/', [
+    (session, args, next) => {
+      qna(session.message.text, (err, result) => {
+        if (err) {
+          console.error(err);
+          session.send('Unfortunately an error occurred. Try again.');
+        }
+        else {
+          session.send(JSON.parse(result).answer);
+        }
+      });
+    }
+  ]);
 }
 
 if (require.main === module) {
