+++ localConfig.json
@@ -1,0 +1,3 @@
+{
+  "qnaUri" : "http://qnaservice.cloudapp.net/KBService.svc/GetAnswer?kbId=a3dae93561d24529b11a4c0eb8acd444&question="
+}
